restructure benches and tests to allow building on stable
